Moved mock ipython fixture to conftest.py
