+++ tests/test_draw.py
@@ -34,16 +34,6 @@ def fig(request):
     plt.close()
 
 
-@pytest.fixture
-def mock_ipython():
-    with patch("IPython.get_ipython") as mock_get_ipython:
-        mock_shell = MagicMock()
-        mock_shell.__class__.__name__ = "ZMQInteractiveShell"
-        mock_shell.config = {"IPKernelApp": True}
-        mock_get_ipython.return_value = mock_shell
-        yield
-
-
 @pytest.mark.parametrize("arg", ("x", 1))
 def test_profile_1(fig, minuit, arg):
     minuit.draw_profile(arg)
